ml_server/requirements.txt
fastapi
uvicorn
torch==2.3.1
torchvision==0.18.1
numpy==1.24.4
pillow
matplotlib
opencv-python
